backend/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "legal-research-backend"
version = "0.1.0"
description = "Legal Research Assistant Backend API"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
langchain = "^0.0.340"
langchain-google-genai = "^0.0.5"
google-generativeai = "^0.3.2"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
aiosqlite = "^0.19.0"
redis = "^5.0.1"
celery = "^5.3.4"
pydantic = "^2.5.0"
httpx = "^0.25.2"
aiofiles = "^23.2.1"
faiss-cpu = "^1.7.4"
tiktoken = "^0.5.2"
pypdf2 = "^3.0.1"
python-docx = "^1.1.0"
sentence-transformers = "^2.2.2"
openai = "^1.3.8"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
